+++ src/main/java/subway/domain/Line.java
@@ -1,120 +1,54 @@
 package subway.domain;
 
-import java.util.ArrayList;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Objects;
-import java.util.stream.Collectors;
 
 public class Line {
 
     private final Long id;
     private final String name;
-    private final LinkedList<Section> sections;
+    private final Sections sections;
 
-    public Line(Long id, String name, LinkedList<Section> sections) {
+    public Line(Long id, String name, Sections sections) {
         this.id = id;
         this.name = name;
         this.sections = sections;
     }
 
     public boolean hasStation(Station station) {
-        return sections.stream()
-                .anyMatch(section -> (section.getLeft().equals(station) || section.getRight().equals(station)));
+        return sections.hasStation(station);
     }
 
     public boolean hasLeftStationInSection(Station station) {
-        return sections.stream()
-                .anyMatch(section -> section.getLeft().equals(station));
+        return sections.hasLeftStationInSection(station);
     }
 
     public boolean hasRightStationInSection(Station station) {
-        return sections.stream()
-                .anyMatch(section -> section.getRight().equals(station));
+        return sections.hasRightStationInSection(station);
     }
 
     public Section findSectionByLeftStation(Station station) {
-        return sections.stream()
-                .filter(section -> section.getLeft().equals(station))
-                .findAny()
-                .orElseThrow(() -> new IllegalArgumentException("일치하는 역이 없습니다."));
+        return sections.findSectionByLeftStation(station);
     }
 
     public Section findSectionByRightStation(Station station) {
-        return sections.stream()
-                .filter(section -> section.getRight().equals(station))
-                .findAny()
-                .orElseThrow(() -> new IllegalArgumentException("일치하는 역이 없습니다."));
+        return sections.findSectionByRightStation(station);
     }
 
     public boolean hasOneSection() {
-        return sections.size() == 1;
+        return sections.hasOneSection();
     }
 
     public boolean isLastStationAtLeft(Station station) {
-        List<Station> leftStations = findLeftStations();
-        List<Station> rightStations = findRightStations();
-
-        leftStations.removeAll(rightStations);
-
-        return leftStations.stream()
-                .filter(lettStation -> lettStation.equals(station))
-                .count() == 1;
+        return sections.isLastStationAtLeft(station);
     }
 
     public boolean isLastStationAtRight(Station station) {
-        List<Station> leftStations = findLeftStations();
-        List<Station> rightStations = findRightStations();
-
-        rightStations.removeAll(leftStations);
-
-        return rightStations.stream()
-                .filter(rightStation -> rightStation.equals(station))
-                .count() == 1;
+        return sections.isLastStationAtRight(station);
     }
 
     public List<Station> findLeftToRightRoute() {
-        if (sections.isEmpty()) {
-            return new ArrayList<>();
-        }
-        Station lastStations = findLastStationAtLeft();
-        List<Station> stations = new ArrayList<>(List.of(lastStations));
-
-        Station targetStation = lastStations;
-        while (stations.size() < sections.size() + 1) {
-            for (Section section : sections) {
-                if (section.getLeft().equals(targetStation)) {
-                    Station nextStation = section.getRight();
-                    stations.add(nextStation);
-                    targetStation = nextStation;
-                    break;
-                }
-            }
-        }
-        return stations;
-    }
-
-    private Station findLastStationAtLeft() {
-        List<Station> leftStations = findLeftStations();
-        List<Station> rightStations = findRightStations();
-
-        leftStations.removeAll(rightStations);
-
-        return leftStations.stream()
-                .findAny()
-                .orElseThrow(() -> new IllegalArgumentException("종점을 찾을 수 없습니다."));
-    }
-
-    private List<Station> findLeftStations() {
-        return sections.stream()
-                .map(Section::getLeft)
-                .collect(Collectors.toList());
-    }
-
-    private List<Station> findRightStations() {
-        return sections.stream()
-                .map(Section::getRight)
-                .collect(Collectors.toList());
+        return sections.findLeftToRightRoute();
     }
 
     public Long getId() {
@@ -126,7 +60,7 @@ public class Line {
     }
 
     public List<Section> getSections() {
-        return sections;
+        return sections.getSections();
     }
 
     @Override
@@ -138,12 +72,11 @@ public class Line {
             return false;
         }
         Line line = (Line) o;
-        return Objects.equals(getId(), line.getId()) && Objects.equals(getName(), line.getName())
-                && Objects.equals(getSections(), line.getSections());
+        return Objects.equals(getId(), line.getId());
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(getId(), getName(), getSections());
+        return Objects.hash(getId());
     }
 }
